Revert decora_wifi cleanup
