+++ mypy.ini
@@ -528,17 +528,6 @@ no_implicit_optional = true
 warn_return_any = true
 warn_unreachable = true
 
-[mypy-homeassistant.components.decora_wifi.*]
-check_untyped_defs = true
-disallow_incomplete_defs = true
-disallow_subclassing_any = true
-disallow_untyped_calls = true
-disallow_untyped_decorators = true
-disallow_untyped_defs = true
-no_implicit_optional = true
-warn_return_any = true
-warn_unreachable = true
-
 [mypy-homeassistant.components.device_automation.*]
 check_untyped_defs = true
 disallow_incomplete_defs = true
